MATTER-4260 Fix Coverity INTEGER_OVERFLOW in crow routing Trie::add
Widen Trie::add() rule_index parameter from uint16_t to unsigned int
to prevent implicit truncation from size_t at the call site (line 1295).
Add explicit static_cast<uint16_t> when assigning to Node::rule_index.

This resolves Coverity CID 53630775 (CWE-190 Integer Overflow).

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)

Reviewed-on: http://androidsource.nxp.com/project/32893

+++ examples/chip-tool/webui-2_0/MatterApplicationBackend/libs/crow/crow/routing.h
@@ -992,7 +992,7 @@ namespace crow // NOTE: Already documented in "crow/app.h"
         }
 
         //This functions assumes any blueprint info passed is valid
-        void add(const std::string& url, uint16_t rule_index, unsigned bp_prefix_length = 0, uint16_t blueprint_index = INVALID_BP_ID)
+        void add(const std::string& url, unsigned int rule_index, unsigned bp_prefix_length = 0, uint16_t blueprint_index = INVALID_BP_ID)
         {
             auto idx = &head_;
 
@@ -1074,7 +1074,7 @@ namespace crow // NOTE: Already documented in "crow/app.h"
             //check if the last node already has a value (exact url already in Trie)
             if (idx->rule_index)
                 throw std::runtime_error("handler already exists for " + url);
-            idx->rule_index = rule_index;
+            idx->rule_index = static_cast<uint16_t>(rule_index);
         }
 
     private:
